requirements.txt
beautifulsoup4==4.6.3
requests==2.20.0
psycopg2-binary==2.8.4
Unidecode==1.0.22
